simple getIntList, where, composition, >>=

+++ evenodds.hs
@@ -1,7 +1,9 @@
 -- Vicfred
 -- https://codeforces.com/problemset/problem/318/A
 
-solve :: [Int] -> Int
+import Data.Int
+
+solve :: [Int64] -> Int64
 solve [n,k] = if k <= m then 2*k-1 else 2*(k-m)
               where m = (n+1) `div` 2
